Amount columns type declaration fixed with unsignedBigInteger (was unsignedInteger)

+++ database/migrations/2024_02_24_164500_create_payu_methods_table.php
@@ -15,8 +15,8 @@ return new class extends Migration {
             $table->boolean('available')->default(false);
             $table->boolean('active')->default(true);
             $table->string('type', 10)->nullable();
-            $table->unsignedInteger('min_amount')->default(0);
-            $table->unsignedInteger('max_amount')->default(99999999);
+            $table->unsignedBigInteger('min_amount')->default(0);
+            $table->unsignedBigInteger('max_amount')->default(99999999);
             $table->timestamps();
         });
     }
